@@ -2,36 +2,49 @@
   "name": "alt-utils",
   "version": "2.0.0",
   "description": "A collection of utils that work with alt.js",
-  "main": "index.js",
+  "main": "lib",
+  "jsnext:main": "src",
   "scripts": {
-    "build": "npm run clean && npm run transpile",
+    "build": "yarn run clean && yarn run transpile",
     "clean": "rimraf lib",
-    "coverage": "npm run transpile-cover && istanbul cover node_modules/mocha/bin/_mocha -- -u exports -R tap --require ./test/babel test",
+    "coverage": "yarn run transpile-cover && babel-node node_modules/mocha/bin/_mocha -- -u exports -R tap test",
     "lint": "eslint src components",
-    "pretest": "npm run clean && npm run transpile",
-    "test": "mocha -u exports -R nyan --require ./test/babel test",
-    "transpile": "babel src --out-dir lib --stage 0",
-    "transpile-cover": "babel src --out-dir lib -r --stage 0"
+    "pretest": "yarn run clean && yarn run transpile",
+    "test": "babel-node node_modules/.bin/_mocha -u exports -R nyan test",
+    "transpile": "babel src --out-dir lib",
+    "transpile-cover": "babel src --out-dir lib --plugins @babel/plugin-external-helpers"
   },
   "author": "Josh Perez <josh@goatslacker.com>",
   "license": "MIT",
   "devDependencies": {
-    "alt": "0.17.4",
-    "babel": "5.8.23",
-    "babel-core": "5.8.25",
-    "chai": "3.3.0",
-    "immutable": "3.7.5",
-    "iso": "4.2.0",
-    "istanbul": "0.3.22",
-    "jsdom": "6.5.1",
-    "mocha": "2.3.3",
-    "react": "0.14.0",
-    "react-addons-test-utils": "0.14.0",
-    "react-dom": "0.14.0",
-    "rimraf": "2.4.3",
-    "sinon": "1.17.1"
+    "@babel/cli": "^7.16.7",
+    "@babel/core": "^7.16.7",
+    "@babel/node": "^7.16.7",
+    "@babel/plugin-external-helpers": "^7.16.7",
+    "@babel/plugin-proposal-decorators": "^7.16.7",
+    "@babel/plugin-proposal-function-bind": "^7.16.7",
+    "@babel/plugin-transform-classes": "^7.16.7",
+    "@babel/plugin-transform-runtime": "^7.16.7",
+    "@babel/preset-env": "^7.16.7",
+    "@babel/preset-react": "^7.16.7",
+    "alt": "git+https://github.com/bitshares/alt.git",
+    "alt-container": "https://github.com/xiangxn/alt-container.git#update-node",
+    "babel-istanbul": "^0.12.2",
+    "babel-plugin-add-module-exports": "^1.0.4",
+    "babel-plugin-istanbul": "^6.1.1",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "chai": "^4.3.4",
+    "immutable": "^4.0.0",
+    "iso": "^5.2.0",
+    "istanbul": "^0.4.5",
+    "jsdom": "^19.0.0",
+    "mocha": "^9.1.3",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "rimraf": "^3.0.2",
+    "sinon": "^12.0.1"
   },
   "dependencies": {
-    "prop-types": "^15.5.10"
+    "prop-types": "^15.8.0"
   }
 }
